requirements.txt
# FastAPI framework and dependencies
fastapi==0.105.0
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # For PostgreSQL support
aiosqlite==0.19.0       # For SQLite async support

# API tools
email-validator==2.1.0
httpx==0.25.2
itsdangerous==2.1.2

# Google OAuth
google-auth==2.23.4
google-auth-oauthlib==1.1.0

# M-Pesa API integration
requests==2.31.0

# Redis for caching and session management
redis==5.0.1

# Task processing
celery==5.3.5

# Image processing
pillow==10.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
isort==5.12.0
mypy==1.7.1